rename metrics dir to details
