+++ pages/reporting/details/index.js
