+++ pages/reporting/details/ppl-sla/content/index.js
