+++ pages/reporting/details/ppl-sla/views/index.jsx
