+++ pages/reporting/details/routes.js
